Synthesized shaft tier (v3.5 Layer C1): kinematic interpolation through P anchors, dark
Per docs/implementation/shaft_position_first_impl_plan.md C1. New
header-only shaft_synthesis.h: between consecutive located P anchors,
theta(t) = monotone-limited (Fritsch-Carlson) cubic Hermite through the
anchor (theta, theta-dot) pairs with the unwrap branch chosen by the
TRACK'S direction of rotation (expectedDelta from the anchor slopes,
not shortest-wrap — a 200-deg rotation interpolates the long way);
grip = per-axis Hermite with pose-path endpoint velocities; length
lerped; conf = min(anchors) decaying to synth.midConfFrac at midspan;
one sample per camera frame STRICTLY between anchors, nothing outside.

ShaftSampleFlags widens uint8 -> uint16 (byte was full) + Synthesized
= 0x100 — the plan's only non-additive type change; all flag-building
casts swept (assembly x3, ball_anchor x2), JSON/QML already int-safe.
club.synth[] serialized on all three paths (swing_doc + toAnalysisDetail
parity, disk_replay_source relTimedMap) — visualization tier, EXCLUDED
from metrics/scoring by flag. Dark: synth.enabled default false;
off ⇒ byte-identical (tested). 40/40 Analysis tests.

+++ src/Analysis/ball_anchor.cpp
@@ -320,7 +320,7 @@ void applyBallAnchor(ShaftTrack2D &out, const BallTrack2D &ball,
             sm.visibleLenPx = std::hypot(ballPx[size_t(i)].x() - gx[size_t(i)],
                                          ballPx[size_t(i)].y() - gy[size_t(i)]);
             sm.conf     = std::max(sm.conf, kAnchorConf);
-            sm.flags    = uint8_t((sm.flags & ~ShaftHeadProjected) | ShaftBallAnchored);
+            sm.flags    = uint16_t((sm.flags & ~ShaftHeadProjected) | ShaftBallAnchored);
         }
     }
 
@@ -344,7 +344,7 @@ void applyBallAnchor(ShaftTrack2D &out, const BallTrack2D &ball,
                 sm.headPx   = QPointF(bx, by);
                 sm.visibleLenPx = std::hypot(bx - gx[size_t(fImp)], by - gy[size_t(fImp)]);
                 sm.conf     = std::max(sm.conf, kAnchorConf);
-                sm.flags    = uint8_t((sm.flags & ~ShaftHeadProjected) | ShaftBallAnchored);
+                sm.flags    = uint16_t((sm.flags & ~ShaftHeadProjected) | ShaftBallAnchored);
             }
         }
     }
